chore: restructure routes
